extracting the captcha webview logic to a reusable class
